requirements.txt
streamlit
Pillow
numpy
librosa
pandas
matplotlib
